feat: add dummy transaction data
This commit introduces a list of dummy transaction data to be used throughout the application. The data is represented as a list of maps, where each map represents a single transaction.

The transaction data includes the following keys:

- 'icon': An FaIcon representing the transaction category (e.g., food, entertainment, etc.)
- 'color': The color associated with the transaction category
- 'name': The name or description of the transaction
- 'totalAmount': The total amount of the transaction, represented as a string
- 'date': The date of the transaction, represented as a string

This dummy data can be used for testing, prototyping, or as a starting point for implementing the actual transaction data handling logic in the application.

Note: This data is for development purposes only and should be replaced with actual data from a proper data source (e.g., API, database) before deploying the application to production.