chore: add font_awesome_flutter and fl_chart dependencies
This commit adds the following dependencies to the project:

- font_awesome_flutter: ^10.7.0
  - This package provides Flutter icons from the Font Awesome icon library, allowing the application to use a wide range of icons for various UI elements.

- fl_chart: ^0.66.2
  - This package provides a highly customizable and interactive charting library for Flutter applications, enabling the creation of various types of charts and visualizations.

These dependencies will be used to enhance the application's user interface and data visualization capabilities. The font_awesome_flutter package will allow for the use of Font Awesome icons throughout the application, while the fl_chart package will enable the creation of charts and graphs to display data in a visually appealing and interactive manner.

By adding these dependencies, the project can leverage the functionality and features provided by these popular and well-maintained packages, improving the overall user experience and visual appeal of the application.